Super WIP everything is very messy
Refactoring cards to just be seperate cards because I realized that they have very different behavior. If commonalities emerge later maybe combine some things.
Added object for en to la translations in order to not accidentally lose the connection between the two. This was probably the appropriate data structure to start with anyway.
There are lots of notes and loose ends about things that could be better now.